Add : Mission4 - 댓글 컴포넌트화, Props로 데이터 전달 구현
